feat(breakout-rooms/native): moved all related components to feature
