+++ react/features/breakout-rooms/components/native/BreakoutRoomNamePrompt.tsx
@@ -1,9 +1,9 @@
 import React, { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 
-import InputDialog from '../../../../../base/dialog/components/native/InputDialog';
-import { renameBreakoutRoom } from '../../../../../breakout-rooms/actions';
-import { IBreakoutRoomNamePromptProps as IProps } from '../../../../types';
+import InputDialog from '../../../base/dialog/components/native/InputDialog';
+import { renameBreakoutRoom } from '../../actions';
+import { IBreakoutRoomNamePromptProps as IProps } from '../../../participants-pane/types';
 
 /**
  * Implements a component to render a breakout room name prompt.
