+++ react/features/breakout-rooms/components/native/CollapsibleRoom.tsx
@@ -3,10 +3,10 @@ import { useTranslation } from 'react-i18next';
 import { FlatList } from 'react-native';
 import { useDispatch } from 'react-redux';
 
-import { openSheet } from '../../../../../base/dialog/actions';
-import { IRoom } from '../../../../../breakout-rooms/types';
-import { participantMatchesSearch } from '../../../../functions';
-import CollapsibleList from '../../../native/CollapsibleList';
+import { openSheet } from '../../../base/dialog/actions';
+import { IRoom } from '../../types';
+import { participantMatchesSearch } from '../../../participants-pane/functions';
+import CollapsibleList from '../../../participants-pane/components/native/CollapsibleList';
 
 import BreakoutRoomContextMenu from './BreakoutRoomContextMenu';
 import BreakoutRoomParticipantItem from './BreakoutRoomParticipantItem';
