+++ react/features/breakout-rooms/components/native/LeaveBreakoutRoomButton.tsx
@@ -1,11 +1,11 @@
 import React, { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 
-import { createBreakoutRoomsEvent } from '../../../../../analytics/AnalyticsEvents';
-import { sendAnalytics } from '../../../../../analytics/functions';
-import Button from '../../../../../base/ui/components/native/Button';
-import { BUTTON_TYPES } from '../../../../../base/ui/constants.native';
-import { moveToRoom } from '../../../../../breakout-rooms/actions';
+import { createBreakoutRoomsEvent } from '../../../analytics/AnalyticsEvents';
+import { sendAnalytics } from '../../../analytics/functions';
+import Button from '../../../base/ui/components/native/Button';
+import { BUTTON_TYPES } from '../../../base/ui/constants.native';
+import { moveToRoom } from '../../actions';
 
 import styles from './styles';
